Updated astro and astro-icon.
This upgrade is a breaking change in `astro-icon` and forced a number of updates.
As a result, this change is significantly larger than I would prefer. For now,
`astro-icon` no longer uses the iconify API to load images. This requires the
installation of the various icon libraries and configuration in the astro.config
to not bloat the install of the unneeded SVGs.