First wave of CSS/SCSS cleanup.
Condensed the hover/focus changes to a single file, making it generic by using
Bootstrap's CSS variables as much as possible. Focusing on the navigation and
dropdowns right now, will focus on the form fields later.

Also, cleaning up a lot of global styles used in the various common components
(Header and Footer) that will be as common as we move to the subsite concept.

The next steps are to move the styles in `bootstrap.scss` that don't need to be
global and should be in individual components, as well as test for any bugs in the
CSS cleanup so far.