SCSS/CSS Cleanup in support of the Escape Room subsites. (#398)
Condensed the hover/focus changes to a single file, making it generic by using
Bootstrap's CSS variables as much as possible. Focusing on the navigation and
dropdowns right now, will focus on the form fields later.

Also, cleaning up a lot of global styles used in the various common components
(Header and Footer) that will be as common as we move to the subsite concept.

Updated the `EscapeRoomLayout` to utilize the common navigation and moved
the prototypes pages to utilize that layout.